Refactor Classes and Update Composer Autoloader
